+++ src/city.h
@@ -5,9 +5,9 @@
 typedef struct
 {
     char* welcomeText;
-    u8 x;
-    u8 y;
+    u16 x;
+    u16 y;
 } City;
 
-City* City_Init(char* name, u8 x, u8 y);
+City* City_Init(char* name, u16 x, u16 y);
 void City_Free(City* city);
